Add files via upload
One of such APIs is the Google Text to Speech API commonly known as the gTTS API.
 gTTS is a very easy to use tool which converts the text entered, into audio which can be saved as a mp3 file.

The gTTS API supports several languages including English, Hindi, Tamil, French, German and many more. 
The speech can be delivered in any one of the two available audio speeds, fast or slow

pip install gTTS



2...
 One of such APIs available in the python library commonly known as win32com library. 
It provides a bunch of methods to get excited about and one of them is the Dispatch method of the library.
 Dispatch method when passed with the argument of SAPI.
SpVoice It interacts with the Microsoft Speech SDK to speak what you type in from the keyboard.

pip install pypiwin32



pyttsx is a cross-platform text to speech library which is platform independent. 
The major advantage of using this library for text-to-speech conversion is that it works offline

pip install pyttsx3